Fix several issues that were causing the proxy to hang (#25)
* Fix hang caused by sending value to unbuffered channel

If internal requests were processed out of order this would lead to a
case where a request response would attempt to send results to
unbuffered channel w/o out a receiver. This would cause that request to
hang indefinitely.

* Add to pending before writing

* Fix race condition dealing with pending requests

The problem is that requests could be added to a client connections
pending requests after it had already been closed. This would cause a
request to never be handled.