Fix prepared queries (#28)
In the existing code execute queries were just a pass-through. That means that
if an unprepared error occurred it was just passed back to the driver. The
problem is that an attempt to prepare that query again would be forwarded to a
connection that was not the one that originally serviced the execute. This
could cause a scenario where execute requests would never resolve from the
unprepared state.

The proxy fixes this by keeping its own cache and properly handling unprepared
errors. This cache was push down into the `proxycore` because eventually we
should add pre-preparing queries on add/up node events. This should be address
in a follow up issue: https://github.com/datastax/cql-proxy/issues/32