Fix the Astra endpoint test (#46)
Fix the Astra endpoint test by creating the mock metadata service once
at the start of the tests. It was failing because listener couldn't bind
the port because it was already in use from the previous test and we
didn't synchronously wait for it to close. It's easier and faster to
share the instance between tests.