Optimize parser
This replaces the existing antlr4 parser with a recursive decent parser
that uses Ragel for the lexer implementation. This results in a
*significant* speed up. A rough benchmark w/ a non-optimal case shows a
~7000x speed up:

ParserAntlr: 187771600 ns/op 0.121691 mib/s
ParserRagel: 24835 ns/op 920.041225 mib/s

A more realistic benchmark wouldn't have as many "handled" queries and
would be faster for the Ragel-based parser.

Currently, this only detects "SELECT" queries as idempotent. A follow up
patch will add that implementation.