Add feature to configure peer proxies
Multiple proxies can be configured to work together for:
* Multi-region failover in Astra DB.
* Bypassing drivers that require at least replication factor number of
  nodes in a cluster for token-aware load balancing.

Also, added a feature to configure the proxy using a YAML configuration
file. This is necessary for configuring multiple proxies for the `peers`
value. All keys match their CLI counterpart:

```yaml
astra-bundle: path/to/secure-connect-bundle.zip
username: <username>
password: <password>
bind: 127.0.0.1:9042
rpc-address: 127.0.0.1
peers:
  - rpc-address: 127.0.0.1
    data-center: dc-1
  - rpc-address: 127.0.0.2
    data-center: dc-
```

`peers` is not available via CLI option because it's a complex,
composite value. It's possible to configure everything on the CLI other
than `peers` and put only that in the configuration file.