Improve listener handling and fix tests
While improving the close handling of the listener in `Proxy` I noticed
that this could be generalized to handle multiple listeners (like
`http.Server`). This could be quite useful in the future for supporting
different multi-proxy features in a more user-friendly and resource
efficient way.  So `Proxy`, the type, now allows multiple listeners in a
single proxy. This hasn't been exposed in an external feature yet.

This also improves clean up for servers, and their listeners, used
during testing and now tests generate unique, unbound ports for each
test to avoid sporadic errors I was seeing in CI.