Implement compression support (#150)
Implement compression support

Compression is negotiated with the client connection then it is
propogated to the upstream coordinator (or Astra) connections and
negotiated. If a frame is not modified then it doesn't need to be
recompressed.

Uses partial message encoding support to rewrite the consistency when
it's not supported. Remove frame patching. This is necessary because the
frames need to be recompressed when they're modified.

---------

Co-authored-by: Michael Penick <michael.penick@datastax.com>